fixes fixes fixes

+++ code/game/gamemodes/endgame/supermatter_cascade/blob.dm
@@ -66,21 +66,19 @@
 		"<span class=\"danger\">You reach out and touch \the [src]. Everything immediately goes quiet. Your last thought is \"That was not a wise decision.\"</span>",\
 		"<span class=\"warning\">You hear an unearthly noise.</span>")
 
-	playsound(src, 'sound/effects/supermatter.ogg', 50, 1)
-
-	Consume(user)
+	if(Consume(user))
+		playsound(src, GET_SFX(SFX_SUPERMATTER), 50, 1)
 
 /turf/unsimulated/wall/supermatter/attackby(obj/item/W as obj, mob/living/user as mob)
 	user.visible_message("<span class=\"warning\">\The [user] touches \a [W] to \the [src] as a silence fills the room...</span>",\
 		"<span class=\"danger\">You touch \the [W] to \the [src] when everything suddenly goes silent.\"</span>\n<span class=\"notice\">\The [W] flashes into dust as you flinch away from \the [src].</span>",\
 		"<span class=\"warning\">Everything suddenly goes silent.</span>")
 
-	playsound(src, 'sound/effects/supermatter.ogg', 50, 1)
-
 	user.drop(W, force = TRUE)
-	Consume(W)
+	if(Consume(W))
+		playsound(src, GET_SFX(SFX_SUPERMATTER), 50, 1)
 
-#define MayConsume(A) (istype(A) && A.simulated && !isobserver(A))
+#define MayConsume(A) (istype(A) && A.simulated && !isobserver(A) && !istype(A, /obj/effect/overlay/bluespacify))
 
 /turf/unsimulated/wall/supermatter/Bumped(atom/movable/AM)
 	if(!MayConsume(AM))
@@ -94,14 +92,15 @@
 		AM.visible_message("<span class=\"warning\">\The [AM] smacks into \the [src] and rapidly flashes to ash.</span>",\
 		"<span class=\"warning\">You hear a loud crack as you are washed with a wave of heat.</span>")
 
-	playsound(src, 'sound/effects/supermatter.ogg', 50, 1)
-	Consume(AM)
+	if(Consume(AM))
+		playsound(src, GET_SFX(SFX_SUPERMATTER), 50, 1)
 
 /turf/unsimulated/wall/supermatter/Entered(atom/movable/AM)
 	Bumped(AM)
 
 /turf/unsimulated/wall/supermatter/proc/Consume(atom/movable/AM)
-	if(MayConsume(AM))
+	if(MayConsume(AM) && !istype(AM, /obj/effect/overlay/bluespacify))
 		qdel(AM)
+		return TRUE
 
 #undef MayConsume
